user_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic
python-jose[cryptography]
passlib>=1.7.4
bcrypt==3.2.0
python-multipart
requests
web3